+++ rollup.plugin.config.js
@@ -1,16 +1,18 @@
-import babel from 'rollup-plugin-babel';
-import replace from 'rollup-plugin-replace';
-import commonjs from 'rollup-plugin-commonjs';
-import nodeResolve from 'rollup-plugin-node-resolve';
+import babel from '@rollup/plugin-babel';
+import replace from '@rollup/plugin-replace';
+import commonjs from '@rollup/plugin-commonjs';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
 import builtins from 'rollup-plugin-node-builtins';
 import globals from 'rollup-plugin-node-globals';
 import eslint from 'rollup-plugin-eslint';
-import sizes from 'rollup-plugin-sizes';
+
 export default {
-  entry: 'lib/index.js',
-  dest: 'dist/bundle.js',
-  format: 'cjs',
-  sourceMap: false,
+  input: 'lib/index.js',
+  output: {
+    file: 'dist/bundle.js',
+    format: 'cjs'
+  },
+  sourcemap: false,
   plugins: [
     eslint({
       exclude: ['src/styles/**']